requirements.txt
# OCR y PDF
pytesseract==0.3.10
pdf2image==1.17.0
PyMuPDF==1.23.24
pdfplumber==0.10.3
Pillow==10.3.0
camelot-py==0.11.0
ghostscript==0.7

# Utilidades
tabulate==0.9.0
whoosh==2.7.4
